Add anacron support
This patch introduces the following changes:
- Add support for batchmode + cron + anacron
- Update alpine version to 3.18

The batchmode + cron/anacron support allows triggering
the backup at regular intervals with full scans, in cases
where changes in the host file system are not detected.

Alpine version has been increased to 3.18. Note that as of
this writing, there is a version conflict between Alpine 3.19
glibc support, and the version of the sqlite provided with
SpiderOakONE client, so Alpine has been pinned to 3.18.

Signed-off-by: Ernest Artiaga <ernest.artiaga@eartiam.net>